Updated payload to reflect Salesforce logic changes
This commit contains updates to the application payload
to reflect logic changes that have been made in the
Salesforce backend. Previously, we were passing both
the new isAlsoApplicant key/value pair, as well as a
mainContactIsAlsoLegalSignatory key/value pair. The
changes that have been made in the Salesforce backend
do not require the extra mainContact key/value pair,
so this has been removed.

Additionally, this commit adds the projectCapitalWorks
field to the payload that is built within buildPayload.ts.
 This not being present was causing local submissions to
fail when hitting Salesforce. Adding it in fixes this error.