Updated dependencies to address security vulnerabilities
GitHub raised a dependency alert against this repository on
the 9th April, reporting that versions of minimist needed to be
bumped to 0.2.1 or later.

I've updated some of the dependencies to address this (more than
one dependency was using minimist, some were using different
versions), which has caused this commit to be quite large.